requirements.txt
Django==4.2.9
django-postgres-extra==2.0.8
djangorestframework==3.14.0
google-api-python-client==2.115.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery[redis]==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0
django-constance==3.1.0
django-dynamic-fixture==4.0.1

